Refactoring OneLoginAuthConnector into two classes to make it more testable
