+++ src/main/java/com/nike/cerberus/server/config/guice/OneLoginGuiceModule.java
@@ -1,0 +1,56 @@
+package com.nike.cerberus.server.config.guice;
+
+import com.fasterxml.jackson.databind.DeserializationFeature;
+import com.fasterxml.jackson.databind.ObjectMapper;
+import com.fasterxml.jackson.databind.PropertyNamingStrategy;
+import com.fasterxml.jackson.databind.SerializationFeature;
+import com.google.inject.AbstractModule;
+import com.google.inject.Provides;
+import okhttp3.OkHttpClient;
+
+import javax.inject.Named;
+import javax.inject.Singleton;
+import java.util.concurrent.TimeUnit;
+
+/**
+ * Guice module for some OneLogin related classes
+ */
+public class OneLoginGuiceModule extends AbstractModule {
+
+    public static final String ONE_LOGIN_OBJECT_MAPPER_NAME = "OneLoginObjectMapper";
+
+    private static final int DEFAULT_TIMEOUT = 15;
+
+    private static final TimeUnit DEFAULT_TIMEOUT_UNIT = TimeUnit.SECONDS;
+
+    @Override
+    protected void configure() {
+    }
+
+    @Provides
+    @Singleton
+    @Named(ONE_LOGIN_OBJECT_MAPPER_NAME)
+    public ObjectMapper getObjectMapper() {
+        ObjectMapper objectMapper = new ObjectMapper();
+        objectMapper.findAndRegisterModules();
+        objectMapper.setPropertyNamingStrategy(PropertyNamingStrategy.SNAKE_CASE);
+        objectMapper.disable(SerializationFeature.WRITE_DATES_AS_TIMESTAMPS);
+        objectMapper.disable(SerializationFeature.FAIL_ON_EMPTY_BEANS);
+        objectMapper.disable(DeserializationFeature.FAIL_ON_UNKNOWN_PROPERTIES);
+        objectMapper.enable(SerializationFeature.WRITE_EMPTY_JSON_ARRAYS);
+        objectMapper.enable(SerializationFeature.INDENT_OUTPUT);
+        return objectMapper;
+    }
+
+    @Provides
+    @Singleton
+    public OkHttpClient getOkHttpClient() {
+        return new OkHttpClient.Builder()
+                .connectTimeout(DEFAULT_TIMEOUT, DEFAULT_TIMEOUT_UNIT)
+                .writeTimeout(DEFAULT_TIMEOUT, DEFAULT_TIMEOUT_UNIT)
+                .readTimeout(DEFAULT_TIMEOUT, DEFAULT_TIMEOUT_UNIT)
+                .build();
+    }
+
+
+}
